Add common base class
Move all Navigation Drawer-related code into a shared base class, to remove the repetition of the same stuff
